using math module
Assume that two input angles A and B given in degrees, complete the function named anglesumtest that will compare cos(A+B) to its equivalent trigonometric identity

cos(A+B)=cos(A)cos(B)-sin(A)sin(B).

The function input variables A and B are angles in degrees. Your function output should be either True or False depending on whether or not the above identity is correct. Your function should perform the following steps:

Convert the input angles from degrees to radians using the math.radians method.
-->  Compute cos(A+B)
-->  Compute cos(A)cos(B)-sin(A)sin(B)
-->  Apply the math.fabs method to compute the absolute value of the difference between the values computed in steps 2 and 3.
If the quantity in step 4 is less than 1e-15, return True; otherwise, return False.